requirements.txt
streamlit
yfinance
pandas
numpy
datetime
ephem
plotly.graph_objects
plotly.express